refactor: extract subcomponents/hooks from ComplianceDashboard
- Extract utility functions into utils.ts
- Extract state management into ComplianceDashboardState.ts
- Extract OverallGaugeComponent from renderOverallGauge
- Extract CategoryBreakdownComponent from renderCategoryBreakdown
- Extract FeatureMatrixComponent from renderFeatureMatrix
- Extract TestStatisticsComponent from renderTestStatistics
- Extract RoadmapTimelineComponent from renderRoadmapTimeline
- Extract RecentUpdatesComponent from renderRecentUpdates
- Extract DatabaseComparisonComponent from renderDatabaseComparison
- Update ComplianceDashboard to use extracted components

This refactoring improves maintainability by separating concerns and
making the code more modular and testable.

Amp-Thread-ID: https://ampcode.com/threads/T-0fe09bb5-c764-4e38-93b6-70287b114652
